% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boost_gaussianLSS.R
\name{boost_gaussianLSS}
\alias{boost_gaussianLSS}
\title{Boosting GaussianLSS Model}
\usage{
boost_gaussianLSS(
  y,
  data,
  m_stop = 1000,
  center_x = TRUE,
  weights = NULL,
  method = "SAASL"
)
}
\arguments{
\item{y}{response variable}

\item{data}{feature matrix}

\item{m_stop}{stopping iteration, default is 1000}

\item{center_x}{center the feature matrix? default is TRUE}

\item{weights}{a vector of weights indicating the training (1) and testing (0) data,
default is NULL, i.e. all observations as training data}

\item{method}{step length method, should be one of {"FSL", "ASL", "SAASL", "SAASL05"},
default is "SAASL"}
}
\value{
\code{mu} the estiamted coefficients of mean \cr
        \code{sigma} the estimated coefficients of the predictor of standard deviation \cr
        \code{mu_mat} matrix of the estimated coefficients of mu in each iteartion \cr
        \code{si_mat} matrix of the estimated coefficients of sigma in each iteration \cr
        \code{v_mu} step length of mu in each iteartion \cr
        \code{v_mu} step length of sigma in each iteration \cr
        \code{muvsi} which distribution paramter is updated in each iteration \cr
        \code{v_mu_var} which variable is used for updating mu in each iteration \cr
        \code{v_si_var} which variable is used for updating sigma in each iteartion \cr
        \code{like_train} positive likelihood of the training data in each iteartion \cr
        \code{like_test} positive likelihood of the test data in each iteration
}
\description{
Fit boosting GaussianLSS model with different kinds of step lengths.
}
