% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_risk.R
\name{cv_risk}
\alias{cv_risk}
\title{Cross-Validation for boosting GaussianLSS}
\usage{
cv_risk(
  y,
  data,
  m_stop = 1000,
  center_x = T,
  folds = cv(rep(1, length(y)), type = "kfold", B = 10),
  method = "SAASL",
  papply = mclapply
)
}
\arguments{
\item{y}{response variable}

\item{data}{feature matrix}

\item{m_stop}{stopping iteration, default is 1000}

\item{center_x}{center the feature matrix? default is TRUE}

\item{folds}{a matrix indicating the folds of CV, specified by mboost::cv,
default is 10-folds CV.}

\item{method}{step length method, should be one of {"FSL", "ASL", "SAASL", "SAASL05"},
default is "SAASL"}

\item{papply}{parallel computation, default is mclapply}
}
\description{
Cross-validation for boosting GaussianLSS
}
