% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_risk.R
\name{cvrisk_plot}
\alias{cvrisk_plot}
\title{cv_risk plot}
\usage{
cvrisk_plot(cv_risk)
}
\arguments{
\item{cv_risk}{likelihood of the cv_risk output}
}
\description{
Plot the negative log-likelihood of the cv_risk output.
}
