% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boost_gaussianLSS.R
\name{getRes}
\alias{getRes}
\title{get residuals and estimated coefficients}
\usage{
getRes(y, X)
}
\arguments{
\item{y}{response variable}

\item{X}{feature matrix}
}
\description{
get residuals and estimated coefficients
}
\keyword{internal}
